Fix CobFile struct data loading
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@213 37977431-3df6-0310-b722-df95706aa16b

+++ rts/CobFile.h
@@ -52,12 +52,12 @@ class CCobFile
 {
 public:
 	vector<string> scriptNames;
-	vector<long> scriptOffsets;
-	vector<long> scriptLengths;			//Assumes that the scripts are sorted by offset in the file
+	vector<int> scriptOffsets;
+	vector<int> scriptLengths;			//Assumes that the scripts are sorted by offset in the file
 	vector<string> pieceNames;
 	vector<int> scriptIndex;
 	map<string, int> scriptMap;
-	long* code;
+	int* code;
 	int numStaticVars;
 	string name;
 public:
